+++ core/lib/zksync_core/src/commitment_generator/metrics.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use vise::{Buckets, EncodeLabelSet, EncodeLabelValue, Family, Histogram, Metrics};
+use vise::{Buckets, EncodeLabelSet, EncodeLabelValue, Family, Histogram, Metrics, Unit};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EncodeLabelValue, EncodeLabelSet)]
 #[metrics(label = "stage", rename_all = "snake_case")]
@@ -15,13 +15,13 @@ pub(super) enum CommitmentStage {
 #[metrics(prefix = "server_commitment_generator")]
 pub(super) struct CommitmentGeneratorMetrics {
     /// Latency of generating commitment per stage.
-    #[metrics(buckets = Buckets::LATENCIES)]
+    #[metrics(buckets = Buckets::LATENCIES, unit = Unit::Seconds)]
     pub generate_commitment_latency_stage: Family<CommitmentStage, Histogram<Duration>>,
     /// Latency of generating bootloader content commitment.
-    #[metrics(buckets = Buckets::LATENCIES)]
+    #[metrics(buckets = Buckets::LATENCIES, unit = Unit::Seconds)]
     pub bootloader_content_commitment_latency: Histogram<Duration>,
     /// Latency of generating events queue commitment.
-    #[metrics(buckets = Buckets::LATENCIES)]
+    #[metrics(buckets = Buckets::LATENCIES, unit = Unit::Seconds)]
     pub events_queue_commitment_latency: Histogram<Duration>,
 }
